refactor: move request related variable and value helper to http module
