+++ chk/modules/http/support.py
@@ -5,6 +5,35 @@ from cerberus.validator import DocumentError
 from chk.infrastructure.exception import err_message
 from chk.modules.http.constants import RequestConfigNode
 from chk.modules.http.validation_rules import request_schema
+from chk.modules.variables.lexicon import StringLexicalAnalyzer
+
+
+class RequestValueHandler:
+    """ Handle variables and values regarding request """
+
+    @staticmethod
+    def request_fill_val(document: dict, symbol_table: dict):
+        """Convert request block variables"""
+        def process_dict(doc: dict, var_s: dict):
+            for key in doc.keys():
+                if type(doc[key]) is str:
+                    item = str(doc[key])
+                    doc[key] = StringLexicalAnalyzer.replace_in_str(item, var_s)
+                elif type(doc[key]) is dict:
+                    doc[key] = process_dict(doc[key], var_s)
+            return doc
+
+        request_document = document.get(RequestConfigNode.ROOT, {})
+        import copy; request_document = copy.deepcopy(request_document)
+
+        return process_dict(request_document, symbol_table)
+
+    @staticmethod
+    def request_get_return(self, document: dict, response: dict) -> dict:
+        if req := document.get(RequestConfigNode.ROOT, {}):
+            if ret := req.get(RequestConfigNode.RETURN):
+                pass  # handle return
+        return {}  # return all
 
 
 class RequestMixin(object):
@@ -30,3 +59,4 @@ class RequestMixin(object):
             return {key:self.document[key] for key in (RequestConfigNode.ROOT, ) if key in self.document}  # type: ignore
         except Exception as ex:
             raise SystemExit(err_message('fatal.V0005', extra=ex))
+
